Update bypass-403.sh
I made some changes that I think are good for the script. First, I added a few payloads, then arranged the order in which they appear in the code to make it easier and neater to add more payloads in the future. Finally, I added some code that makes the script display its usage and terminate execution if no parameters were passed. :)